backend/requirements.txt
flask
flask-cors
sqlalchemy
psycopg2-binary
pandas
openpyxl
requests
python-dotenv